Added validator module and renamed parse/serialise modules
