+++ src/serialiser.js
@@ -85,13 +85,6 @@ class Serialiser {
   result = {}
 }
 
-function serialise (data, ast) {
-  const serialiser = new Serialiser(data, ast)
-
-  return serialiser.serialise()
-}
-
 module.exports = {
-  Serialiser,
-  serialise
+  Serialiser
 }
